Attempt to handle escaping and wrapping around inline markup
It's a bit tricky to nicely handle the boundaries of inline markup when
formatting and wrapping a paragraph: you have to know when `\ ` is
necessary and make sure not to break in the middle of one, even though
it's a space. (You could add `\ ` around markup unconditionally, but
that would look pretty silly.) Here's an attempt at doing that.