Tweak parser settings to disable includes and raise more errors
Configure the parser so that warnings (e.g., malformed tables) cause the
parser to raise an exception rather than print a warning and return an
empty document; otherwise, we take input and return nothing. Also
disable the `include` directive, which should not be processed in this
setting, and add our own handling for it.