Update testing code to account for Black application
Applying Black to a block of Python may change the spacing in it, which
will lead to a different AST at the reST level and hence to a failed
comparison. Instead, we now have a special case for checking Python code
blocks that applies Black to both sides and compares the results.

An alternative might be to use the `ast` module for parsing, but that
always uses the grammar of the current version of Python, which would
cause the behavior of this tool to be tied more closely to the version
of Python.