Switch to using overlined characters for the top two header levels
Following the Python Developer's Guide, this switches to using `#` and
`*` (overlined) for the first two header levels. The use of overlined
headers can look a bit loud, but, after staring at a lot of reST source,
I'm coming around to the opinion that having more visual distinction
between the different levels is for the best. Plus, this means the more
weird characters don't have to be used as much.