Add Mypy type annotations
The annotations aren't perfect yet, due to some subtleties about whether
the formatting of a node returns lines or inline markup depending on the
type of the node, but they should be helpful.

`__init__.py` is added so that `mypy rstfmt` works (evidently the
`namespace_packages` option doesn't help there).