Refactor and comment reST extensions module
I was trying to go for a thing where the module set up data structures
describing the constructs to add and `register` just iterated over that,
but that ended up being messier than just having the logic in
`register`. This also means that less logic happens during import, which
is what I really wanted anyway, and makes it easier to add more
directives elsewhere.

This also changes directive parsing so that directives added here can
take any argument without causing a parse error. A formatting tool
shouldn't fail because the name of the argument to a function is
mistyped.