Improve handling of roles: support aliases and more roles
This allows, for example, ":pep-reference:`3000`" and equivalent roles
to be rewritten to ":PEP:`3000`". The role class that
`register_generic_role` uses internally doesn't seem to expose the role
that was actually used nicely; that could be extracted from the raw
source, but it seemed easier to just have a new role to allow us to
manage the role name and body independently and not have to reparse
things.

Also add support for some more roles, including all standard docutils
roles.