Remove `--in-place` option and default to writing back files
I was thinking of sed with the `-i`/`--in-place` option, but this
matches Black (and rustfmt) better. While gofmt works more like the
previous version, with its `-w`, I figure Black is the closest analog to
rstfmt, thanks to the Python associations of both, and concatenating
multiple formatted files to stdout isn't that useful a behavior anyway.