Handle manually enumerated lists
rstfmt now handles enumerated lists that have a starting number
explicitly specified. Only the number on the first item is preserved;
nobody wants to worry about manually keeping numbers in sync, I
imagine. The parse tree doesn't seem to distinguish auto-enumerated
lists from ones with an explicit first number of 1, so both of those are
rendered as auto-enumerated lists for compatibility (and because I think
that looks better in the common case of all auto-enumerated lists).