Handle substitution definitions
The addition of `docutils.nodes.Inline` as a base class of
`rst_extras.directive` is necessitated by a spot in the parsing code for
substitution definitions where it filters out nodes that aren't
instances of `Inline` or `Text`. I somewhat fear that this will cause
unintended consequences somewhere down the line, but it appears to be
working so far. The only alternative I can see is to do some horrible
monkey patching of the substitution definition class; I'd like to avoid
that if at all possible.