Add partial support for Sphinx glossary directive
The glossary directive is annoying in that it contains custom parsing
logic to accept an extended version of the standard definition list
syntax (allowing multiple terms to be attached to one definition and
allowing inline markup in terms). Dealing with that fully will require
considerably more work, but this change at least allows glossary
directives to be handled at all, as long as their bodies are valid as
standard reStructuredText.

An alternative would be to simply treat them as raw directives, but I
think that would be more of a loss than a gain, considering that this
should cover most cases and the starting point is not accepting the
directive at all.