Account for removal of `rawsource` attribute on `Text` nodes
The attribute was removed in 0.18; per the release notes [1], we should
instead use the string value of the node itself (`Text` is a subclass of
`str`). That value comes from `docutils.utils.escape2null`, which
replaces some backslashes with null characters [2], so we need to
reverse that when working from the string value.

And now that we work with the latest release, change the docutils
requirement to specify a minimum instead, in line with the other ones.

[1] https://docutils.sourceforge.io/RELEASE-NOTES.html#release-0-18-2021-10-26

[2] If I understand the code correctly, it replaces all backslashes
    except those that are escaped by other backslashes (with those ones
    then not counting as escaping any immediately following
    backslashes). But it doesn't really matter; all we need to know to
    undo it is that all it can do is replace backslashes with nulls.